fix bug with duplication of default cities on first launch

+++ app/src/main/java/by/intervale/akella266/weather2/api/models/Clouds.java
@@ -7,21 +7,21 @@ public class Clouds {
 
     @SerializedName("all")
     @Expose
-    private Integer all;
+    private Double all;
 
     public Clouds() {
     }
 
-    public Clouds(Integer all) {
+    public Clouds(Double all) {
         super();
         this.all = all;
     }
 
-    public Integer getAll() {
+    public Double getAll() {
         return all;
     }
 
-    public void setAll(Integer all) {
+    public void setAll(Double all) {
         this.all = all;
     }
